Shifted the Home to redesigned Dashboard
Shifted the Home to redesigned Dashboard. Users will be see the this new Dashboard from now. Also media loading is now added to HomeActivity. In addition to that, removed the animation of auto hid bottom navigation bar and search bar while scrolling for simplicity. It may come back once everything else is settled.